+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 - Added better error checking for CLI commands and support for mutually
 exclusive fields.
 - Added `--interactive` flag to CLI which is required to run interactive mode.
+- Added CLI role rule-add Organization and Environment interactive prompts.
 
 ### Changed
 - Silenced `begin` supports human readable time (Format: Jan 02 2006 3:04PM MST)